requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
Django==3.2
django-redis==5.2.0
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7
redis==4.4.0
sqlparse==0.4.3
